Refactor encryption/decryption algorithm of the actual message blob to a hybrid algorithm using AES and RSA
A randomly generated AES key will be used to encrypt the message blob,
where AES key will be encrypted with a public RSA key and is
included as part of the final encryption blob. The decryption
algorithm uses the AWS KMS service to decrypt the private part of
the RSA key which will be used to decrypt the AES key too finally
decrypt the actual message blob with it.